+++ relm4/src/component/async/traits.rs
@@ -41,6 +41,13 @@ pub trait AsyncComponent: Sized + 'static {
     /// Initializes the root widget
     fn init_root() -> Self::Root;
 
+    /// Allows you to initialize the root widget with a temporary value
+    /// as a placeholder until the [`init()`](AsyncComponent::init)
+    /// future completes.
+    ///
+    /// This method does nothing by default.
+    fn init_loading_widgets(_root: &mut Self::Root) {}
+
     /// Creates the initial model and view, docking it into the component.
     async fn init(
         init: Self::Init,
@@ -145,6 +152,13 @@ pub trait SimpleAsyncComponent: Sized + 'static {
     /// Initializes the root widget
     fn init_root() -> Self::Root;
 
+    /// Allows you to initialize the root widget with a temporary value
+    /// as a placeholder until the [`init()`](AsyncComponent::init)
+    /// future completes.
+    ///
+    /// This method does nothing by default.
+    fn init_loading_widgets(_root: &mut Self::Root) {}
+
     /// Creates the initial model and view, docking it into the component.
     async fn init(
         init: Self::Init,
@@ -184,6 +198,10 @@ where
 
     fn init_root() -> Self::Root {
         C::init_root()
+    }
+
+    fn init_loading_widgets(root: &mut Self::Root) {
+        C::init_loading_widgets(root)
     }
 
     async fn init(