+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### Added
 
++ core: Temporary widget initialization for `AsyncComponent`s
 + core: Add asynchronous factories including macro support
 + core: Add `Reducer` as message based alternative to `SharedState`
 + core: Runtimes of `Component`s and `AsyncComponents` can now be detached for a static lifetime